chore: delete non nessesary files and move new files for client and server
