+++ 05_remote_procedure_call/src/client/method_table.js
