+++ 05_remote_procedure_call/src/server/method_table.py
@@ -1,27 +1,29 @@
-from .functions import floor, nroot, reverse, validAnagram, sort
-from typing import List, Dict, Any
+from functions import floor, nroot, reverse, validAnagram, sort
 
-FunctionTable = Dict[str, Dict[str, Any]]
-
-method_table: FunctionTable = {
-    "floor" : {
-        "function" : floor,
-        "paramTypes" : [float]
+method_table = {
+    "floor": {
+        "function": floor,
+        "param_types": [ float ],
+        "result_type": "int"
+    }, 
+    "nroot": {
+        "function": nroot,
+        "param_types": [ int, int ],
+        "result_type": "int"
     },
-    "nroot" : {
-        "function" : nroot,
-        "paramTypes" : [int, int]
+    "reverse": {
+        "function": reverse,
+        "param_types": [ str ],
+        "result_type": "str"
     },
-    "reverse" : {
-        "function" : reverse,
-        "paramTypes" : [str]
+    "validAnagram": {
+        "function": validAnagram,
+        "param_types": [ str, str ],
+        "result_type": "bool"
     },
-    "validAnagram" : {
-        "function" : validAnagram,
-        "paramTypes" : [str, str]
-    },
-    "sort" : {
-        "function" : sort,
-        "paramTypes" : [List[str]]
+    "sort": {
+        "function": sort,
+        "param_types": [ list[str] ],  
+        "result_type": "list[str]"
     }
 }
